Windows: fix various compile issues
* syntax error typo from SBuf conversion of service_name

* add several missing includes

* fix macro wrapper around mswsock.h

* in ipc_win2.cc fix parameter error for memset() and several potential
  buffer overruns by removing magic number parameters on buffer size.

* fdd_table entry is not a HANDLE but a function pointer

+++ src/fde.h
@@ -166,7 +166,7 @@ private:
         dynamicSslContext = NULL;
 #endif
 #if _SQUID_WINDOWS_
-        win32.handle = NULL;
+        win32.handle = (long)NULL;
 #endif
         tosFromServer = '\0';
         nfmarkFromServer = 0;
